flashcard_app/backend/requirements.txt
fastapi
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
uvicorn
python-multipart